other problems solved
The majority of issues pertaining to the pup_unpacker.py have been successfully addressed. However, there persists a lingering concern concerning the proper recognition of the MAGIC attribute within PUP files, thus rendering them invalid. My conjecture is that the root of this quandary resides within the code of the ps4_dec_pup_info.py file, which has not been brought up to date with the most recent PS4 firmware iteration. Should this script undergo a comprehensive update, it could potentially facilitate the extraction of the latest PUP files for both the PS4 and PS5 platforms.