fixes and improvements
I have fixed the annoying crash that occurred during log writing by adding a timestamp between reading the logs from the terminal and writing them to the interface. I also implemented the display of PUP file information directly on the GUI. To view this information, simply enlarge the window or set the application to full screen. You can find the compiled application in the "Release" section.